test: restructure tests
split up tests

- unit
- api
- wallets
and
regtest
