+++ Makefile
@@ -6,6 +6,8 @@ format: prettier black ruff
 
 check: mypy pyright checkblack checkruff checkprettier checkbundle
 
+test: test-unit test-wallets test-api test-regtest
+
 prettier:
 	poetry run ./node_modules/.bin/prettier --write lnbits
 
@@ -36,19 +38,28 @@ checkeditorconfig:
 dev:
 	poetry run lnbits --reload
 
-test:
+test-wallets:
 	LNBITS_BACKEND_WALLET_CLASS="FakeWallet" \
-	FAKE_WALLET_SECRET="ToTheMoon1" \
-	LNBITS_DATA_FOLDER="./tests/data" \
 	PYTHONUNBUFFERED=1 \
 	DEBUG=true \
-	poetry run pytest
+	poetry run pytest tests/wallets
 
-test-real-wallet:
-	LNBITS_DATA_FOLDER="./tests/data" \
+test-unit:
+	LNBITS_BACKEND_WALLET_CLASS="FakeWallet" \
 	PYTHONUNBUFFERED=1 \
 	DEBUG=true \
-	poetry run pytest
+	poetry run pytest tests/unit
+
+test-api:
+	LNBITS_BACKEND_WALLET_CLASS="FakeWallet" \
+	PYTHONUNBUFFERED=1 \
+	DEBUG=true \
+	poetry run pytest tests/api
+
+test-regtest:
+	PYTHONUNBUFFERED=1 \
+	DEBUG=true \
+	poetry run pytest tests/regtest
 
 test-migration:
 	LNBITS_ADMIN_UI=True \
